Lab4 IntegrationPatterns/code/RPI_gateway_CoAP/requirements.txt
CoAPthon3~=1.0.2